+++ tests/testthat/test-spgautor.R
@@ -259,6 +259,7 @@ test_that("generics work spgautor polygon data with missing", {
   expect_true(inherits(predict(spmod1, interval = "confidence", level = 0.9), "matrix"))
   expect_true(inherits(predict(spmod1, type = "terms"), "matrix"))
   expect_type(predict(spmod1, type = "terms", interval = "confidence"), "list")
+  expect_vector(predict(spmod1, dispersion = 1))
 
   # print
   expect_output(print(spmod1))