+++ tests/testthat/test-spglm.R
@@ -122,6 +122,7 @@ test_that("generics work spglm point data", {
   expect_true(inherits(predict(spmod1, newdata = newexdata, type = "response", interval = "confidence", level = 0.9), "matrix"))
   expect_true(inherits(predict(spmod1, newdata = newexdata, type = "terms"), "matrix"))
   expect_type(predict(spmod1, newdata = newexdata, type = "terms", interval = "confidence"), "list")
+  expect_vector(predict(spmod1, newdata = newexdata, dispersion = 1))
 
   # print
   expect_output(print(spmod1))
@@ -284,6 +285,7 @@ test_that("generics work spglm point data with missing", {
   expect_true(inherits(predict(spmod1, newdata = newexdata, type = "response", interval = "confidence", level = 0.9), "matrix"))
   expect_true(inherits(predict(spmod1, newdata = newexdata, type = "terms"), "matrix"))
   expect_type(predict(spmod1, newdata = newexdata, type = "terms", interval = "confidence"), "list")
+  expect_vector(predict(spmod1, newdata = newexdata, dispersion = 1))
 
   # print
   expect_output(print(spmod1))
@@ -441,6 +443,7 @@ test_that("generics work spglm polygon data with missing", {
   expect_true(inherits(predict(spmod1, type = "response", interval = "confidence", level = 0.9), "matrix"))
   expect_true(inherits(predict(spmod1, type = "terms"), "matrix"))
   expect_type(predict(spmod1, type = "terms", interval = "confidence"), "list")
+  expect_vector(predict(spmod1, dispersion = 1))
 
   # print
   expect_output(print(spmod1))